Move 2 non-integration tests to their rightful places
