+++ src/NBitcoin.Tests/ConstructorsForSerializableClassesTest.cs
@@ -1,11 +1,9 @@
 ﻿using System;
 using System.Collections.Generic;
-using System.IO;
 using System.Linq;
-using NBitcoin;
 using Xunit;
 
-namespace Stratis.Bitcoin.IntegrationTests
+namespace NBitcoin.Tests
 {
     public class ConstructorsForSerializableClassesTest
     {
@@ -16,13 +14,13 @@ namespace Stratis.Bitcoin.IntegrationTests
             // parameterless constructor for a good reason so they should not fail this test.
             List<Type> exceptionalTypes = new List<Type>()
             {
-                typeof(NBitcoin.ExtKey),
-                typeof(NBitcoin.ExtPubKey),
-                typeof(NBitcoin.PubKey),
-                typeof(NBitcoin.PosBlock),
-                typeof(NBitcoin.PosBlockHeader),
-                typeof(NBitcoin.PosTransaction),
-                typeof(NBitcoin.Protocol.CompactVarInt),
+                typeof(ExtKey),
+                typeof(ExtPubKey),
+                typeof(PubKey),
+                typeof(PosBlock),
+                typeof(PosBlockHeader),
+                typeof(PosTransaction),
+                typeof(Protocol.CompactVarInt),
             };
 
             var types = AppDomain.CurrentDomain.GetAssemblies()
@@ -40,24 +38,6 @@ namespace Stratis.Bitcoin.IntegrationTests
                 Assert.True(parameterlessConstructorExists, $"Class {type.FullName} inherits {typeof(IBitcoinSerializable).Name} " +
                     "but doesn't have a parameterless constructor which is needed for serialization and deserialization process!");
             }
-        }
-
-        /// <summary>
-        /// Serializes input and returns deserialized object.
-        /// </summary>
-        /// <remarks>Needed for troubleshooting this test.</remarks>
-        private T CloneViaSerializeDeserialize<T>(T input) where T : IBitcoinSerializable
-        {
-            MemoryStream ms = new MemoryStream();
-            BitcoinStream bitcoinStream = new BitcoinStream(ms, true);
-
-            input.ReadWrite(bitcoinStream);
-            ms.Position = 0;
-
-            bitcoinStream = new BitcoinStream(bitcoinStream.Inner, false);
-            var obj = Activator.CreateInstance<T>();
-            obj.ReadWrite(bitcoinStream);
-            return obj;
         }
     }
 }
